+++ src/Movement/Move2.cpp
@@ -408,7 +408,7 @@ GCodeResult Move::ConfigureNonlinearExtrusion(GCodeBuffer& gb, const StringRef& 
 // Return the position of an endstop in steps. Only called for kinematics that home drives individually.
 int32_t Move::GetEndstopPositionSteps(size_t drive, bool highEnd) noexcept
 {
-	return kinematics->GetEndstopPosition(drive, highEnd) * driveStepsPerMm[drive];
+	return (int32_t)(kinematics->GetEndstopPosition(drive, highEnd) * driveStepsPerMm[drive]);
 }
 
 // This is called from the step ISR as well as other places, so keep it fast
